+++ search-films/src/layouts/RequireAuth/index.tsx
@@ -1,10 +1,11 @@
-import { ReactNode, useContext } from 'react';
+import { ReactNode } from 'react';
+import { useSelector } from 'react-redux';
 import { Navigate } from 'react-router-dom';
-import { IUserContext, UserContext } from '../../context/user.context';
+import { RootState } from '../../store/store';
 
 export const RequireAuth = ( { children } : {children: ReactNode} ) => {
-	const { isAuthenticated } = useContext(UserContext) as IUserContext;
-	if(!isAuthenticated) {
+	const currentUser = useSelector((s: RootState) => s.users.currentUser);
+	if(!currentUser) {
 		return <Navigate to="/login" replace/>;
 	} else {
 		return children;